check for existence of .vimrc before backing it up

+++ install.sh
@@ -1,3 +1,6 @@
+if [ -f ~/.vimrc ]
+then 
 cp ~/.vimrc ~/.vimrc.bak
+fi
 cp -r .vim ~/.vim
 cp .vimrc ~/.vimrc
